Iframe support in the dom (#11)

+++ src/admin.py
@@ -9,6 +9,7 @@ from flask import Blueprint, flash, redirect, render_template, request
 from data import User, PR, add_pr, fix_date
 from forms import PRForm, ModifyPRForm
 from werkzeug.utils import secure_filename
+from urllib.parse import quote
 
 admin_page = Blueprint("admin", __name__)
 
@@ -40,8 +41,10 @@ def admin():
 
     if form.validate_on_submit():
         filename = form.file.data.filename
-        if not filename or not allowed_file(filename):
-            flash("File type not supported")
+        link = form.link.data
+
+        if link and filename:
+            flash("Both file and link")
             return redirect("/admin")
 
         # Check if start date is after end date
@@ -57,20 +60,43 @@ def admin():
                 flash(msg)
                 return redirect("/admin")
 
-        org_filename = secure_filename(filename)
-        
-        # Generate random filename with correct extention
-        filename = str(uuid.uuid4()) + "." + \
-            org_filename.rsplit('.', 1)[1].lower()
-        form.file.data.save(os.path.join(
-            app.config['UPLOAD_FOLDER'], filename))
-        add_pr(file_name=filename,
-               desc=form.desc.data,
-               priority=form.priority.data,
-               start_date=form.start_date.data,
-               end_date=form.end_date.data,
-               user_id=current_user.id,
-               owner=current_user.username)
+        if link:
+            if not link.startswith("https://"):
+                flash("Link didn't start with https://")
+                return redirect("/admin")
+
+            if link.startswith("https://tv.dtek.se"):
+                flash("No links to tv-mannnen D:")
+                return redirect("/admin")
+
+            add_pr(file_name=link,
+                is_iframe=True,
+                desc=form.desc.data,
+                priority=form.priority.data,
+                start_date=form.start_date.data,
+                end_date=form.end_date.data,
+                user_id=current_user.id,
+                owner=current_user.username)
+        elif filename:
+            if (not filename or not allowed_file(filename)):
+                flash("File type not supported")
+                return redirect("/admin")
+
+            org_filename = secure_filename(filename)
+
+            # Generate random filename with correct extention
+            filename = str(uuid.uuid4()) + "." + \
+                org_filename.rsplit('.', 1)[1].lower()
+            form.file.data.save(os.path.join(
+                app.config['UPLOAD_FOLDER'], filename))
+            add_pr(file_name=filename,
+                is_iframe=False,
+                desc=form.desc.data,
+                priority=form.priority.data,
+                start_date=form.start_date.data,
+                end_date=form.end_date.data,
+                user_id=current_user.id,
+                owner=current_user.username)
         return redirect("/admin")
     else:
         # Change the default start and end dates
@@ -107,7 +133,8 @@ def delete():
         return redirect("/admin")
 
     try:
-        os.remove(os.path.join(config.UPLOAD_FOLDER, pr.file_name))
+        if not pr.is_iframe:
+            os.remove(os.path.join(config.UPLOAD_FOLDER, pr.file_name))
     except: 
         flash("PR wasn't found on disk but the database entry has been removed")
